Move some classes around to bust up a package cycle.
