+++ subprojects/kotlin-dsl-integ-tests/src/integTest/kotlin/org/gradle/kotlin/dsl/integration/ProjectSchemaAccessorsIntegrationTest.kt
@@ -840,43 +840,30 @@ class ProjectSchemaAccessorsIntegrationTest : AbstractPluginIntegrationTest() {
 
             class MyPlugin : Plugin<Project> {
                 override fun apply(project: Project): Unit = project.run {
+                    val rootExtension = extensions.create("rootExtension", MyExtension::class.java, "root")
 
-                    val instantiator = serviceOf<Instantiator>()
-
-                    val rootExtension = MyExtension("root", instantiator)
-                    val rootExtensionNestedExtension = MyExtension("nested-in-extension", instantiator)
-                    val rootExtensionNestedConvention = MyConvention("nested-in-extension", instantiator)
-
-                    extensions.add("rootExtension", rootExtension)
-
-                    rootExtension.extensions.add("nestedExtension", rootExtensionNestedExtension)
+                    val rootExtensionNestedExtension = rootExtension.extensions.create("nestedExtension", MyExtension::class.java, "nested-in-extension")
                     rootExtensionNestedExtension.extensions.add("deepExtension", listOf("foo", "bar"))
 
+                    val rootExtensionNestedConvention = objects.newInstance(MyConvention::class.java, "nested-in-extension")
                     rootExtensionNestedConvention.extensions.add("deepExtension", mapOf("foo" to "bar"))
 
-                    val rootConvention = MyConvention("root", instantiator)
-                    val rootConventionNestedExtension = MyExtension("nested-in-convention", instantiator)
-                    val rootConventionNestedConvention = MyConvention("nested-in-convention", instantiator)
+                    val rootConvention = objects.newInstance(MyConvention::class.java, "root")
+                    val rootConventionNestedConvention = objects.newInstance(MyConvention::class.java, "nested-in-convention")
 
                     convention.plugins.put("rootConvention", rootConvention)
 
-                    rootConvention.extensions.add("nestedExtension", rootConventionNestedExtension)
+                    val rootConventionNestedExtension = rootConvention.extensions.create("nestedExtension", MyExtension::class.java, "nested-in-convention")
                     rootConventionNestedExtension.extensions.add("deepExtension", listOf("bazar", "cathedral"))
 
                     rootConventionNestedConvention.extensions.add("deepExtension", mapOf("bazar" to "cathedral"))
                 }
             }
 
-            class MyExtension(val value: String = "value", instantiator: Instantiator) : ExtensionAware, HasConvention {
-                private val convention: DefaultConvention = DefaultConvention(instantiator)
-                override fun getExtensions(): ExtensionContainer = convention
-                override fun getConvention(): Convention = convention
+            abstract class MyExtension(val value: String) : ExtensionAware {
             }
 
-            class MyConvention(val value: String = "value", instantiator: Instantiator) : ExtensionAware, HasConvention {
-                private val convention: DefaultConvention = DefaultConvention(instantiator)
-                override fun getExtensions(): ExtensionContainer = convention
-                override fun getConvention(): Convention = convention
+            abstract class MyConvention @javax.inject.Inject constructor(val value: String) : ExtensionAware {
             }
         """)
