+++ subprojects/model-core/src/main/java/org/gradle/internal/instantiation/generator/ConstructorSelector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.instantiation;
+package org.gradle.internal.instantiation.generator;
 
 /**
  * Encapsulates the differences, due to backwards compatibility, in instantiation for several different types.
