+++ subprojects/model-core/src/main/java/org/gradle/internal/instantiation/generator/DependencyInjectingInstantiator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.instantiation;
+package org.gradle.internal.instantiation.generator;
 
 import org.gradle.api.Describable;
 import org.gradle.api.reflect.ObjectInstantiationException;
+import org.gradle.internal.instantiation.InstanceFactory;
+import org.gradle.internal.instantiation.InstanceGenerator;
 import org.gradle.internal.logging.text.TreeFormatter;
 import org.gradle.internal.reflect.Instantiator;
 import org.gradle.internal.reflect.JavaReflectionUtil;
