+++ subprojects/model-core/src/main/java/org/gradle/internal/instantiation/generator/Jsr330ConstructorSelector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.instantiation;
+package org.gradle.internal.instantiation.generator;
 
 import org.gradle.api.Transformer;
 import org.gradle.cache.internal.CrossBuildInMemoryCache;
