+++ subprojects/model-core/src/test/groovy/org/gradle/internal/instantiation/generator/AbstractClassGeneratorTestGroovy.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,26 @@
  * limitations under the License.
  */
 
+package org.gradle.internal.instantiation.generator
 
+class AbstractClassGeneratorTestGroovy {
 
-package org.gradle.internal.instantiation
+    public static class BeanWithGroovyBoolean {
+        boolean smallB
+        Boolean bigB
+        Boolean mixedB
 
-class TestDecoratedGroovyBean {
-    def String prop
+        boolean getMixedB() { mixedB }
 
-    def doStuff(String value) {
-        "{$value}"
+        Boolean isMixedB() { mixedB }
+
+        void setMixedB(Boolean mixedB) {
+            this.mixedB = mixedB
+        }
+
+        void setMixedB(boolean mixedB) {
+            this.mixedB = mixedB
+        }
     }
+
 }
