+++ subprojects/model-core/src/test/groovy/org/gradle/internal/instantiation/generator/AsmBackedClassGeneratorTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.internal.instantiation;
+package org.gradle.internal.instantiation.generator;
 
 import com.google.common.base.Joiner;
 import groovy.lang.Closure;
@@ -48,6 +48,8 @@ import org.gradle.cache.internal.TestCrossBuildInMemoryCacheFactory;
 import org.gradle.internal.extensibility.ConventionAwareHelper;
 import org.gradle.internal.extensibility.ExtensibleDynamicObject;
 import org.gradle.internal.extensibility.NoConventionMapping;
+import org.gradle.internal.instantiation.ClassGenerationException;
+import org.gradle.internal.instantiation.InstantiatorFactory;
 import org.gradle.internal.metaobject.BeanDynamicObject;
 import org.gradle.internal.metaobject.DynamicObject;
 import org.gradle.internal.reflect.DirectInstantiator;
@@ -80,7 +82,7 @@ import java.util.Set;
 import java.util.concurrent.Callable;
 
 import static org.gradle.api.reflect.TypeOf.typeOf;
-import static org.gradle.internal.instantiation.AbstractClassGeneratorTestGroovy.BeanWithGroovyBoolean;
+import static org.gradle.internal.instantiation.generator.AbstractClassGeneratorTestGroovy.BeanWithGroovyBoolean;
 import static org.gradle.util.Matchers.isEmpty;
 import static org.gradle.util.TestUtil.TEST_CLOSURE;
 import static org.gradle.util.TestUtil.call;
