+++ subprojects/model-core/src/test/groovy/org/gradle/internal/instantiation/generator/TestDecoratedGroovyBean.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,14 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.instantiation;
 
-class PackageScopedClass {
+
+package org.gradle.internal.instantiation.generator
+
+class TestDecoratedGroovyBean {
+    def String prop
+
+    def doStuff(String value) {
+        "{$value}"
+    }
 }
