+++ subprojects/model-core/src/test/groovy/org/gradle/internal/instantiation/generator/TestJavaObject.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,26 +14,10 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.instantiation
+package org.gradle.internal.instantiation.generator;
 
-class AbstractClassGeneratorTestGroovy {
-
-    public static class BeanWithGroovyBoolean {
-        boolean smallB
-        Boolean bigB
-        Boolean mixedB
-
-        boolean getMixedB() { mixedB }
-
-        Boolean isMixedB() { mixedB }
-
-        void setMixedB(Boolean mixedB) {
-            this.mixedB = mixedB
-        }
-
-        void setMixedB(boolean mixedB) {
-            this.mixedB = mixedB
-        }
+public class TestJavaObject {
+    public String getSomeValue() {
+        return "value";
     }
-
 }
